Fixes and simplifies project specific Maven commands
The README.md defines several Maven commands to perform several actions
(e.g. building the application, building the application with WildFly
10, starting the application, etc.).

The command for building the application using WildFly 10 was not
working due to a bad dependencies configuration and to a missing profile
in the command.

This commit fixes this error. In addition, the Arquillian dependency
version has been updated and the README.md commands have been
simplified.