Makes interaction calculus asynchronous
Different species interactions calculus is expected to last some
minutes. This makes necessary to perform calculation in an asynchronous
manner, avoiding max request timeout problems and making users wait for
their results with the client opened.

This commit prepares the application to perform interaction calculus
asynchronously. Although same species interaction calculus should be
almost immediate, it was also modified to be performed asynchronously,
as the different species calculation will be done.

As a result of the changes done, the interactions queries now return a
"work" object with information about the asynchronous work done
(interaction calculation). This work object provides information about
the state of the calculus and the location of the results.