Undoes interactome REST entities id change
The interactome id in REST entities was changed in previous commits to
include the interactome identifier and URL. This was not coherent with
the other REST entities, that only returned the identifier as the id.

This commit undoes this change, and now the interactome returns only its
identifier as the id.