Adds support for different species interactions query
Interaction queries for different species can now be done. This is a big
commit, as the calculus of different species interactions is a complex
process that involves BLAST execution, in this case, through a Docker
machine.

One important change that introduces this commit is the inclusion of
several configuration parameters in the standalone.xml files, that
should be set carefully.