Fixes Docker file path generation
The path of the files in the Docker container are generated with Java's Path
class. This class uses the host's file separator, which may not be the same as
the used in the container system.

This commit fixes this problem by replacing the host's file separator with the
one required by the container system before mounting the volumes.