Fixes wrong sequence order in gene data
The order of the sequences of a gene is important, as each sequence is
identified by the gene it belongs to and its index. However, the gene
data returned by the REST API didn't ordered the sequences, making them
impossible to identify.

This commit fixes this problem and now the Gene entity returns the
sequences in order.