Modifies result's FASTA generation
The FASTA generated for the target interactome comprised the genes of
the target interactome that appears in any interaction. This caused some
of the genes that appeared in the BLAST result not to be returned in the
FASTA. In this commit, this behavior has been changed and now all the
target genes in the BLAST results are returned in the FASTA file.

In addition, a new REST resource has been added to get a FASTA file with
all the genes in a result (in any interactome of the result). This is
more interesting for same genes queries, where all the genes in the
FASTA will belong to the same species.

Finally, the media type of the FASTA REST resources now is
"text/x-fasta".