Adds support for multiple degrees in an interaction
The degree on which a relation may appear in an interaction depends on
the queried interactome. This means that, for example, in a same species
query the same interaction may appear in a different degree on the two
queried interactomes.

Currently, the interactions returned in the results only contained a
single degree. Now, each interactome in each interaction may have a
different degree.