Redesigns database for better interaction data consistency
Due to the current database design, it was possible to create
interactions between genes that do not belong to the same species or
that belong to a species different from their interactome's species.

As such cases are incorrect, this commit modifies the database design to
avoid them. Now, it is required for a gene to belong to the same species
as the interactome to which the interactions where it participates
belong.