Replaces work and result auto-incremental ids
Work and result resources are publicly available. Using an
auto-incremental id for this resources makes it easy for users to guess
the URL of works and results that they didn't request.

These identifiers have been replaced by UUIDs, that can be used safely
to identify these resources and avoid users to access other users'
results.