Redesigns work entities
The work entity has been converted into an abstract entity and the same
and different species results are now subclasses of the work entity.

This change avoids the creation of two different identifiers for the
same calculus (one for the work and another for the result) and
simplifies the database.