Adds a resource for retrieving filtered interactions
Interactions when the degree is 3 may be too much (several tens of
thousands) and it is not practical to return them all in a single
request response.

Therefore, this commit adds an specific resource for retrieving the
iteractions in several parts (paginated). In addition, the interactions
result can now be requested with the "summarize=true" parameter to avoid
including the interactions in the response.