Fixes interaction listing error
Due to an error, all the interactions in the database were returned for
every interaction result. The cause was a missing "where" condition to
filter the interactions by result id.

This commit fixes this error and also introduces a minor performance
tweak, using the InteractionsResult.getInteractions() method instead of
a custom SQL query when no InteractionGroupResultListingOptions is
configured.