Fixes different species interactions algorithm
The fact that a genes can have multiple orthologs after BLAST causes
some strange results. For example, when a query gene as an ortholog and
this ortholog has multiple orthologs it is possible to get interactions
without the query gene appear, in queries with maximum degree 1.

To solve this problem, the algorithm has been limited, and now only
interactions from the reference interactome/s are returned. That is, an
interaction for two orthologs (target interactome/s) that doesn't appear
in reference interactomes will not be returned.