Avoids exceptions in interactions results linkage
The interactions result linkage actions thrown an exception when any of
the requested results where not available because, for example, it was
deleted or it already belongs to another user. This was too strict, and
now the operations return a list of successfully and unsuccessfully
linked results instead of throwing an exception.