Fix count in listing methods
This commit fixes the value returned for the header count in
listing methods that use filtering. To do so, the ListingOption
object used for listing is also passed to the count methods, so
that the total number of elements that meet the filter criteria
are returned.

This commit also fixes the corresponding count method in the
DAOHelper, which was not applying the filters.