Improves species listing and allows FASTA export
This commit improves species listing by supporting sorting,
pagination and filtering (by species name). Also, a method for
downloading the species FASTA with the associated genes is
added. For this, the GeneStorageService was modified to store
the species FASTA in a temporary directory to avoid re-creating
them in different requests.