Redesign interactome entities to support new types
This commit redesigns the interactome entities to support new types. So far,
there was only an Interactome entity that could be associated to one species.
Now, we require to associate an interactome entity with two species, meaning
that there can be inter-species interactions. In addition, we need to store
predictomes, which are like predictomes but must be kept separated because
their interactions are predictions instead of actual interactions obtained
by carrying out the corresponding experiments.

Thus, the Interactome entity now has two sub-entities: DatabaseInteractome
and Predictome. In addition, the Interactome entity has two species and,
consequently, the Interaction has them too.

Although everything is working and it is possible to store the new database,
it is neccessary to test the frontend to make some issueus that may arise
due to the new entities introduced (e.g. the DefaultInteractomeMapper only
works if all interactomes are DatabaseInteractome entities).