Reduces gene queries when retrieving interaction results
We noticet that the DefaultInteractionsMapper was generating a lot of
queries to retrieve the information about genes involved in all the
interactions of a given result. To reduce this queries, this commit
removes "`GeneNamesData[]" creation in the "DefaultInteractionsMapper"
and adds gene names to the InteractionResultData. This is enough for
the frontend visualization requirements.